Initial Upload
Initial upload of files associated with the AGNOSTIC manuscript. These includes python scripts to simulate data, create basis sets, generate the synthetic data, randomize the examples to use during training, build the out-of-voxel artifacts, and train the deep learning models. The best performing models (hdf5) for OOV Detection and OOV Prediction that are shown in the manuscript are included.